fix(list): 单元测试 .jsx -> .tsx
